Struct gat_lending_iterator::Filter

pub struct Filter<I, P> { /* private fields */ }

A lending iterator that filters the elements of iter with predicate.

This struct is created by the filter method on LendingIterator. See its documentation for more.